Increased MaxSpace int size (#80)
* Increased MaxSpace int size

This change prevents errors when python smtplib calls hydroxide.

* Updated UsedSpace to int64

To prevent errors when numbers exceed 32 bit capacity.

+++ protonmail/users.go
@@ -7,10 +7,10 @@ import (
 type User struct {
 	ID         string
 	Name       string
-	UsedSpace  int
+	UsedSpace  int64
 	Currency   string // e.g. EUR
 	Credit     int
-	MaxSpace   int
+	MaxSpace   int64
 	MaxUpload  int
 	Role       int // TODO
 	Private    int
